More seperation using partial definitons of Vector. Concat updated to use LRU and Flags - needs Testing
