+++ DataScience/Core/Vector/AccessRow.cs
@@ -1,0 +1,14 @@
+﻿
+
+namespace DataScience
+{
+    public partial class Vector
+    {
+
+        public static Vector AccessRow(Vector vector, int row)
+        {
+            return new Vector(vector.gpu, vector.Value[(row * vector.Columns)..((row + 1) * vector.Columns)], 1);
+        }
+
+    }
+}
